refactor(website): use new metadata api instead of `head.tsx` (#9269)
