+++ apps/website/src/app/docs/packages/[package]/[version]/[item]/page.tsx
@@ -3,16 +3,23 @@ import { readFile } from 'node:fs/promises';
 import { join } from 'node:path';
 // eslint-disable-next-line n/prefer-global/process
 import process, { cwd } from 'node:process';
-import { createApiModel } from '@discordjs/scripts';
+import { createApiModel, tryResolveSummaryText } from '@discordjs/scripts';
 import type {
 	ApiClass,
+	ApiDeclaredItem,
 	ApiEnum,
 	ApiInterface,
 	ApiItem,
+	ApiItemContainerMixin,
+	ApiMethod,
+	ApiMethodSignature,
+	ApiProperty,
+	ApiPropertySignature,
 	ApiTypeAlias,
 	ApiVariable,
 } from '@microsoft/api-extractor-model';
-import { ApiFunction } from '@microsoft/api-extractor-model';
+import { ApiItemKind, ApiFunction } from '@microsoft/api-extractor-model';
+import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 import { fetchModelJSON } from '~/app/docAPI';
 import { Class } from '~/components/model/Class';
@@ -28,6 +35,79 @@ export interface ItemRouteParams {
 	item: string;
 	package: string;
 	version: string;
+}
+
+async function fetchHeadMember({ package: packageName, version, item }: ItemRouteParams): Promise<ApiItem | undefined> {
+	const modelJSON = await fetchModelJSON(packageName, version);
+	const model = createApiModel(modelJSON);
+	const pkg = model.tryGetPackageByName(packageName);
+	const entry = pkg?.entryPoints[0];
+
+	if (!entry) {
+		return undefined;
+	}
+
+	const [memberName] = decodeURIComponent(item).split(OVERLOAD_SEPARATOR);
+
+	return findMember(model, packageName, memberName);
+}
+
+function resolveMemberSearchParams(packageName: string, member: ApiItem): URLSearchParams {
+	const params = new URLSearchParams({
+		pkg: packageName,
+		kind: member?.kind,
+		name: member?.displayName,
+	});
+
+	switch (member?.kind) {
+		case ApiItemKind.Interface:
+		case ApiItemKind.Class: {
+			const typedMember = member as ApiItemContainerMixin;
+
+			const properties = typedMember.members.filter((member) =>
+				[ApiItemKind.Property, ApiItemKind.PropertySignature].includes(member.kind),
+			) as (ApiProperty | ApiPropertySignature)[];
+			const methods = typedMember.members.filter((member) =>
+				[ApiItemKind.Method, ApiItemKind.Method].includes(member.kind),
+			) as (ApiMethod | ApiMethodSignature)[];
+
+			params.append('methods', methods.length.toString());
+			params.append('props', properties.length.toString());
+			break;
+		}
+
+		case ApiItemKind.Enum: {
+			const typedMember = member as ApiEnum;
+			params.append('members', typedMember.members.length.toString());
+			break;
+		}
+
+		default:
+			break;
+	}
+
+	return params;
+}
+
+export async function generateMetadata({ params }: { params: ItemRouteParams }): Promise<Metadata> {
+	const member = (await fetchHeadMember(params))!;
+	const name = `discord.js${member?.displayName ? ` | ${member.displayName}` : ''}`;
+	const ogTitle = `${params.package ?? 'discord.js'}${member?.displayName ? ` | ${member.displayName}` : ''}`;
+	const url = new URL('https://discordjs.dev/api/og_model');
+	const searchParams = resolveMemberSearchParams(params.package, member);
+	url.search = searchParams.toString();
+	const ogImage = url.toString();
+	const description = tryResolveSummaryText(member as ApiDeclaredItem);
+
+	return {
+		title: name,
+		description: description ?? 'Discord.js API Documentation',
+		openGraph: {
+			title: ogTitle,
+			description: description ?? 'Discord.js API Documentation',
+			images: ogImage,
+		},
+	};
 }
 
 export async function generateStaticParams({ params: { package: packageName, version } }: { params: ItemRouteParams }) {
